overshield complete
Can be picked up ONLY when overshield is not already full. There is a maximum overshield. Please check prefab matches what you want. default is set to 50 for testing purposes. Left it that way so we don't immediately die.

Worth noting there is no mesh on it. as of right now it is a white sphere